chore: extract shared formatBytes utility, remove 3 duplicate copies
